requirements.txt
numpy
PyYAML
nibabel
tqdm
torch
torchvision
